@@ -36,11 +36,11 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  go_router: ^6.0.1
+  go_router: ^9.0.0
   flutter_typeahead: ^4.1.0
   anim_search_bar: ^2.0.2
   flutter_dotenv: ^5.0.2
-  flex_color_scheme: ^7.0.4
+  flex_color_scheme: ^7.1.2
   bloc: ^8.1.0
   flutter_bloc: ^8.1.1
   equatable: ^2.0.5
@@ -67,28 +67,29 @@ dependencies:
   confetti: ^0.7.0
   material_floating_search_bar: ^0.3.7
   url_launcher: ^6.1.6
-  intl: ^0.17.0
+  intl: ^0.18.0
   shared_preferences: ^2.0.15
-  google_sign_in: ^6.0.1
-  image_picker: ^0.8.6
+  google_sign_in: ^6.1.4
+  image_picker: ^1.0.0
   firebase_storage: ^11.0.5
   smooth_page_indicator: ^1.0.0+2
   firebase_crashlytics: ^3.0.6
-  purchases_flutter: ^4.4.1
+  purchases_flutter: ^5.2.1
   roundcheckbox: ^2.0.5
   blobs: ^2.0.0
   settings_ui: ^2.0.2
   showcaseview: ^2.0.0+1
-  flutter_iconpicker:
-    git:
-      url: https://github.com/flutterWithChris/FlutterIconPicker
+  flutter_iconpicker: ^3.2.4
   fluttertoast: ^8.2.1
-  firebase_ui_auth: ^1.2.2
-  firebase_ui_oauth_apple: ^1.1.2
-  firebase_ui_oauth_google: ^1.1.2
-  firebase_ui_oauth: ^1.2.2
-  sign_in_with_apple: ^4.3.0
+  firebase_ui_auth: ^1.1.17
+  firebase_ui_oauth_apple: ^1.0.24
+  firebase_ui_oauth_google: ^1.0.21
+  firebase_ui_oauth: ^1.1.17
+  sign_in_with_apple: ^5.0.0
   permission_handler: ^10.2.0
+  easy_refresh: ^3.3.1
+  flutter_gutter: ^0.1.2
+  chat_gpt_sdk: ^2.2.2
 
       
 dev_dependencies:
