@@ -20,16 +20,15 @@ import 'package:loading_animation_widget/loading_animation_widget.dart';
 class CategoryPageAppBar extends StatefulWidget {
   const CategoryPageAppBar({
     Key? key,
-    required this.contributors,
-    required this.placeList,
+    this.contributors,
+    this.placeList,
     required this.scrollController,
-    required this.listOwner,
+    this.listOwner,
   }) : super(key: key);
 
   final List<User>? contributors;
-
-  final PlaceList placeList;
-  final User listOwner;
+  final PlaceList? placeList;
+  final User? listOwner;
   final ScrollController scrollController;
 
   @override
@@ -38,6 +37,7 @@ class CategoryPageAppBar extends StatefulWidget {
 
 class _CategoryPageAppBarState extends State<CategoryPageAppBar> {
   EdgeInsets avatarStackPadding = const EdgeInsets.only(right: 4.0);
+  EdgeInsets iconPadding = const EdgeInsets.only(right: 4.0, left: 4.0);
 
   @override
   void initState() {
@@ -50,7 +50,7 @@ class _CategoryPageAppBarState extends State<CategoryPageAppBar> {
               Platform.isIOS
                   ? avatarStackPadding =
                       const EdgeInsets.symmetric(horizontal: 28.0)
-                  : avatarStackPadding = const EdgeInsets.only(right: 28.0);
+                  : avatarStackPadding = const EdgeInsets.only(right: 0.0);
             });
           } else if (widget.scrollController.hasClients &&
               widget.scrollController.offset < 65) {
@@ -70,7 +70,7 @@ class _CategoryPageAppBarState extends State<CategoryPageAppBar> {
   Widget build(BuildContext context) {
     return SliverAppBar.medium(
       leadingWidth: 50,
-      expandedHeight: 125,
+      expandedHeight: 120,
       // leading: Padding(
       //   padding: const EdgeInsets.symmetric(horizontal: 8.0),
       //   child: IconButton(
@@ -89,112 +89,133 @@ class _CategoryPageAppBarState extends State<CategoryPageAppBar> {
         child: Row(
           mainAxisAlignment: MainAxisAlignment.spaceBetween,
           children: [
-            Flexible(
-              flex: 4,
-              child: FittedBox(
-                child: Wrap(
-                  spacing: 16.0,
-                  crossAxisAlignment: WrapCrossAlignment.center,
-                  children: [
-                    Animate(
-                        effects: const [
-                          FadeEffect(),
-                          SlideEffect(
-                              duration: Duration(milliseconds: 400),
-                              curve: Curves.easeOutSine,
-                              begin: Offset(-1.0, 0),
-                              end: Offset(0, 0))
-                          // RotateEffect(
-                          //     curve: Curves.easeOutBack,
-                          //     duration: Duration(milliseconds: 500),
-                          //     begin: -0.25,
-                          //     end: 0.0,
-                          //     alignment: Alignment.centerLeft)
-                        ],
-                        child: Padding(
-                          padding: const EdgeInsets.only(right: 4.0, left: 4.0),
-                          child: SizedBox(
+            BlocBuilder<SavedPlacesBloc, SavedPlacesState>(
+              builder: (context, state) {
+                if (state is SavedPlacesLoading) {
+                  return Padding(
+                    padding: const EdgeInsets.only(left: 8.0),
+                    child: LoadingAnimationWidget.beat(
+                        color: FlexColor.bahamaBlueDarkSecondary, size: 20.0),
+                  );
+                }
+                if (state is SavedPlacesLoaded) {
+                  return Flexible(
+                    flex: 4,
+                    child: FittedBox(
+                      child: Wrap(
+                        spacing: 16.0,
+                        crossAxisAlignment: WrapCrossAlignment.center,
+                        children: [
+                          AnimatedPadding(
+                            duration: const Duration(milliseconds: 300),
+                            padding:
+                                const EdgeInsets.only(right: 4.0, left: 4.0),
+                            child: SizedBox(
                               width: 30,
                               height: 30,
-                              child: Icon(
-                                deserializeIcon(widget.placeList.icon) ??
-                                    Icons.list_alt_rounded,
-                                color: Theme.of(context).brightness ==
-                                        Brightness.dark
-                                    ? Colors.white
-                                    : Colors.grey[800],
-                                size: widget.placeList.icon
-                                        .containsValue('fontAwesomeIcons')
-                                    ? 28
-                                    : 32,
-                              )),
-                        )),
-                    Animate(
-                      effects: const [
-                        FadeEffect(),
-                        RotateEffect(
+                              child: context
+                                          .watch<SavedPlacesBloc>()
+                                          .state
+                                          .placeList
+                                          ?.icon !=
+                                      null
+                                  ? Icon(
+                                      deserializeIcon(context
+                                              .read<SavedPlacesBloc>()
+                                              .state
+                                              .placeList!
+                                              .icon) ??
+                                          Icons.list_alt_rounded,
+                                      color: Theme.of(context).brightness ==
+                                              Brightness.dark
+                                          ? Colors.white
+                                          : Colors.grey[800],
+                                      size: context
+                                              .read<SavedPlacesBloc>()
+                                              .state
+                                              .placeList!
+                                              .icon
+                                              .containsValue('fontAwesomeIcons')
+                                          ? 28
+                                          : 32,
+                                    )
+                                  : const Icon(
+                                      Icons.list_alt_rounded,
+                                      color: Colors.white,
+                                      size: 32,
+                                    ),
+                            ),
+                          ),
+                          context
+                                      .watch<SavedPlacesBloc>()
+                                      .state
+                                      .placeList
+                                      ?.name !=
+                                  null
+                              ? Text(
+                                  context
+                                      .read<SavedPlacesBloc>()
+                                      .state
+                                      .placeList!
+                                      .name,
+                                  style: Theme.of(context)
+                                      .textTheme
+                                      .titleLarge!
+                                      .copyWith(fontWeight: FontWeight.bold),
+                                )
+                              : const SizedBox.shrink(),
+                        ],
+                      ).animate().fadeIn().rotate(
                             curve: Curves.easeOutBack,
-                            duration: Duration(milliseconds: 500),
+                            duration: const Duration(milliseconds: 500),
                             begin: -0.25,
                             end: 0.0,
-                            alignment: Alignment.centerLeft)
-                      ],
-                      child: Text(
-                        widget.placeList.name,
-                        style: Theme.of(context)
-                            .textTheme
-                            .titleLarge!
-                            .copyWith(fontWeight: FontWeight.bold),
-                      ),
+                            alignment: Alignment.centerLeft,
+                          ),
                     ),
-                  ],
-                ),
-              ),
+                  );
+                } else {
+                  return const SizedBox();
+                }
+              },
             ),
             Flexible(
-              child: Animate(
-                effects: const [
-                  FadeEffect(),
-                  SlideEffect(
-                      duration: Duration(milliseconds: 700),
-                      curve: Curves.easeOutQuart,
-                      begin: Offset(0.5, 0.0),
-                      end: Offset(0.0, 0.0)),
-                ],
-                child: BlocConsumer<SavedPlacesBloc, SavedPlacesState>(
-                  listener: (context, state) {
-                    if (state is PurchasesUpdated) {
-                      Navigator.pop(context);
-                    }
-                  },
-                  builder: (context, state) {
-                    if (state is SavedPlacesLoading) {
-                      return LoadingAnimationWidget.beat(
-                          color: FlexColor.bahamaBlueDarkSecondary, size: 18.0);
-                    }
-                    if (state is SavedPlacesLoaded) {
-                      return AvatarStack(
-                        settings: RestrictedPositions(
-                            align: StackAlign.right, laying: StackLaying.first),
-                        borderWidth: 2.0,
-                        borderColor:
-                            Theme.of(context).chipTheme.backgroundColor,
-                        width: 70,
-                        height: 40,
-                        avatars: [
-                          CachedNetworkImageProvider(
-                              widget.listOwner.profilePicture!),
-                          for (User user in state.contributors)
-                            CachedNetworkImageProvider(user.profilePicture!),
-                        ],
-                      );
-                    } else {
-                      return const Center(
-                        child: Text('Error Loading Avatars!'),
-                      );
-                    }
-                  },
-                ),
+              child: BlocConsumer<SavedPlacesBloc, SavedPlacesState>(
+                listener: (context, state) {
+                  if (state is PurchasesUpdated) {
+                    Navigator.pop(context);
+                  }
+                },
+                builder: (context, state) {
+                  if (state is SavedPlacesLoading) {
+                    return const SizedBox();
+                  }
+                  if (state is SavedPlacesLoaded) {
+                    return AvatarStack(
+                      settings: RestrictedPositions(
+                          align: StackAlign.right, laying: StackLaying.first),
+                      borderWidth: 2.0,
+                      borderColor: Theme.of(context).chipTheme.backgroundColor,
+                      width: 70,
+                      height: 40,
+                      avatars: [
+                        CachedNetworkImageProvider(
+                            state.listOwner.profilePicture!),
+                        for (User user in state.contributors)
+                          CachedNetworkImageProvider(user.profilePicture!),
+                      ],
+                    ).animate().fadeIn().slideX(
+                          curve: Curves.easeOutSine,
+                          duration: const Duration(milliseconds: 400),
+                          begin: 0.5,
+                          end: 0.0,
+                        );
+                  } else {
+                    return const Center(
+                      child: Text('Error Loading Avatars!'),
+                    );
+                  }
+                },
               ),
             ),
           ],
@@ -216,7 +237,7 @@ class _CategoryPageAppBarState extends State<CategoryPageAppBar> {
                       await showDialog(
                         context: context,
                         builder: (context) => EditListDialog(
-                          placeList: widget.placeList,
+                          placeList: widget.placeList!,
                         ),
                       ).then((value) async {
                         if (value == true) {
@@ -228,9 +249,9 @@ class _CategoryPageAppBarState extends State<CategoryPageAppBar> {
                         }
                       });
                     }),
-                    child: Row(
+                    child: const Row(
                       mainAxisSize: MainAxisSize.min,
-                      children: const [
+                      children: [
                         Icon(Icons.edit_note_rounded),
                         SizedBox(
                           width: 4.0,
@@ -247,15 +268,15 @@ class _CategoryPageAppBarState extends State<CategoryPageAppBar> {
                           context: context,
                           builder: (context) {
                             return DeleteListDialog(
-                              placeList: widget.placeList,
+                              placeList: widget.placeList!,
                             );
                           },
                         );
                       });
                     },
-                    child: Row(
+                    child: const Row(
                       mainAxisSize: MainAxisSize.min,
-                      children: const [
+                      children: [
                         Icon(Icons.delete_forever_rounded),
                         SizedBox(
                           width: 4.0,
